requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
pyinstrument==4.6.2
pyroscope-io==0.8.7
snakeviz==2.2.0
tornado==6.4.1
Werkzeug==3.0.3
